Sentiment Analysis
This file shows an example of using sentiment analysis to understand the lyrics of a song as positive, neutral or negative.  This is just a fun example that doesn't require a lot of experience with sentiment analysis.  We will be using a Bayesian algorithm, but the slides I upload will give more information on other methods (like Recursive Neural Networks).  Enjoy and have fun!